+++ common/api_adapter.js
