+++ index.js
@@ -4,9 +4,9 @@
  */
 
 const corsMiddleware = require('cors')();
-const { martha_v1_handler } = require('./martha_v1');
-const { martha_v2_handler } = require('./martha_v2');
-const { fileSummaryV1Handler } = require('./fileSummaryV1');
+const { martha_v1_handler } = require('./martha_v1/martha_v1');
+const { martha_v2_handler } = require('./martha_v2/martha_v2');
+const { fileSummaryV1Handler } = require('./fileSummaryV1/fileSummaryV1');
 
 exports.martha_v1 = (req, res) => {
     corsMiddleware(req, res, () => martha_v1_handler(req, res));
