+++ martha_v2/martha_v2.js
@@ -1,5 +1,5 @@
-const helpers = require('./helpers');
-const apiAdapter = require('./api_adapter');
+const helpers = require('../common/helpers');
+const apiAdapter = require('../common/api_adapter');
 const URL = require('url');
 
 const BondProviders = Object.freeze({
