+++ test/fileSummaryV1.test.js
@@ -9,10 +9,10 @@
 
 const test = require('ava');
 const sinon = require('sinon');
-const fileSummaryV1 = require('../fileSummaryV1').fileSummaryV1Handler;
-const saKeys = require('../service_account_keys');
-const metadataApi = require('../metadata_api');
-const urlSigner = require('../urlSigner');
+const fileSummaryV1 = require('../fileSummaryV1/fileSummaryV1').fileSummaryV1Handler;
+const saKeys = require('../fileSummaryV1/service_account_keys');
+const metadataApi = require('../fileSummaryV1/metadata_api');
+const urlSigner = require('../fileSummaryV1/urlSigner');
 
 const mockRequest = (req) => {
     req.method = 'POST';
